doc(readme): Add language in code section (#1703) [skip ci]
* doc(readme): Add language in code section

In the whole of DigitalService we talked about our default for using German and English mixed in the code. We are already doing this hybrid approach in our codebase thoroughly. Mostly, we identified on a Digitalservice level, the risk that open-source contributors might not feel invited when they see a lot of German in the variable names.
Therefore, we add an explanation in the readme for everyone to see.

This will probably be used as a DigitalService-wide base, so very happy about your feedback!

* doc(readme): Adapt wording + formatting